+++ somni-expr/CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
 
 - The TypeSet trait has been reworked for better ergonomics. The parser's TypeSet is now an associated type. Additional associated types and trait bounds have been added.
+- The `MemoryRepr` trait has been removed
 
 # [0.1.0] - 2025-07-24